+++ client/src/Components/Dashboard/Proposals/ProposalStyling.jsx
@@ -1,0 +1,24 @@
+import { makeStyles } from "@material-ui/core/styles";
+
+const useStyles = makeStyles({
+    comp: {
+      display: "flex",
+      marginLeft: -60,
+      paddingBottom: 10,
+      justifyContent: "center",
+      alignItem: "center"
+    },
+    multicomp: {
+      display: "block",
+    },
+    index: {
+      display: "flex"
+    },
+    pro: {
+      marginTop: 70,
+      marginLeft: 170,
+      fontSize: "24px"
+    },
+});
+
+export default useStyles();
