+++ client/src/Components/Dashboard/Proposals/SelectedProposals.jsx
@@ -1,52 +1,32 @@
 import React from "react";
 import Header from "../../Layout/Header";
 import Profile from "../DashboardProfile";
-import { makeStyles } from "@material-ui/core/styles";
+import useStyles from "./ProposalStyling";
 import PropDetail from "./ProposalDetail";
-
-const useStyles = makeStyles({
-  comp: {
-    display: "flex",
-    marginLeft: -60,
-    justifyContent: "center",
-    alignItem: "center"
-  },
-  multicomp: {
-    display: "block",
-  },
-  index: {
-    display: "flex"
-  },
-  pro: {
-    marginTop: 70,
-    marginLeft: 170,
-    fontSize: "24px"
-  },
-});
 
 export default function Detail() {
   const classes = useStyles();
   return (
     <div>
       <Header name="Dashboard" />
-        <div className={classes.comp}>
-            <div className={classes.multicomp}>
-                <div className={classes.pro}>
-                    <p>Profile</p>
-                </div>
-                <Profile />
-            </div>
-            <div className={classes.multicomp}>
-                <div className={classes.pro}>
-                    <p>Selected Projects</p>
-                </div>
-                <PropDetail />
-                <div className={classes.pro}>
-                    <p>Other Proposals</p>
-                </div>
-                <PropDetail />
-            </div>
-        </div>
+      <div className={classes.comp}>
+          <div className={classes.multicomp}>
+              <div className={classes.pro}>
+                <p>Profile</p>
+              </div>
+              <Profile />
+          </div>
+          <div className={classes.multicomp}>
+              <div className={classes.pro}>
+                <p>Selected Projects</p>
+              </div>
+              <PropDetail />
+              <div className={classes.pro}>
+                <p>Other Proposals</p>
+              </div>
+              <PropDetail />
+          </div>
+      </div>
     </div>
   );
 }
